organize components, and properly import public images
