+++ src/components/Rgbirdflop/Decode.tsx
