+++ src/components/Rgbirdflop/Output.tsx
