+++ src/components/docs/Breadcrumbs.tsx
@@ -1,6 +1,6 @@
 import { component$ } from '@builder.io/qwik';
 import { Link, useLocation } from '@builder.io/qwik-city';
-import { createBreadcrumbs } from '~/components/docs/SideBar';
+import { createBreadcrumbs } from '~/components/Docs/SideBar';
 import { buildMenu } from '~/util/buildMenu';
 import { useMarkdownItems } from '~/routes/docs/layout';
 import { ChevronRight, Home } from 'lucide-icons-qwik';
