+++ src/components/home/ExpensesChart.tsx
@@ -1,6 +1,6 @@
 // components/elements/HistoricLinePlot.tsx
 import { component$ } from '@builder.io/qwik';
-import Chart from '~/components/Chart';
+import Chart from '~/components/Elements/Chart';
 
 export default component$(() => {
   const config = {
