+++ src/routes/docs/FAQ/reimbursement/index.mdx
@@ -7,7 +7,7 @@ author: Oli/bwmp
 contributors:
   - bwmp
 ---
-import HistoricLinePlot from '~/components/home/HistoricLinePlot';
+import HistoricLinePlot from '~/components/Home/HistoricLinePlot';
 
 # Client Reimbursements
